+++ firedrake/netgen.py
@@ -6,8 +6,10 @@ This file was copied from ngsPETSc.
 import numpy as np
 from scipy.spatial.distance import cdist
 
-from pyop2.mpi import COMM_WORLD
+from pyop2.mpi import COMM_WORLD, MPI
 from firedrake.petsc import PETSc
+from firedrake.cython.dmcommon import DistributedMeshOverlapType
+from firedrake.utils import IntType
 import firedrake
 
 # Netgen and ngsPETSc are not available when the documentation is getting built
@@ -77,6 +79,299 @@ def netgen_distribute(V: firedrake.functionspaceimpl.WithGeometryBase,
             plex_data[(..., *i)] = arr
         plex_data = plex_data.reshape(-1, *nshape[1:])
     return plex_data
+
+
+def adaptive_netgen_cells(mesh):
+    tdim = mesh.topological_dimension
+    if tdim == 2:
+        return mesh.netgen_mesh.Elements2D()
+    elif tdim == 3:
+        return mesh.netgen_mesh.Elements3D()
+    raise NotImplementedError("Adaptive refinement is only implemented in dimension 2 and 3.")
+
+
+def adaptive_netgen_cell_count(mesh):
+    return getattr(mesh, "_adaptive_netgen_num_cells", len(adaptive_netgen_cells(mesh)))
+
+
+def _adaptive_netgen_cell_data(mesh, values):
+    from firedrake.functionspace import FunctionSpace
+
+    DG0 = FunctionSpace(mesh, "DG", 0)
+    data = np.asarray(netgen_distribute(DG0, values), dtype=IntType)
+    cstart, cend = mesh.topology_dm.getHeightStratum(0)
+    cell_numbers = np.fromiter(
+        (mesh._cell_numbering.getOffset(cell) for cell in range(cstart, cend)),
+        dtype=IntType,
+        count=cend - cstart,
+    )
+    ncell = min(data.shape[0], cell_numbers.size)
+    result = np.full(mesh.cell_set.total_size, -1, dtype=IntType)
+    valid = (0 <= cell_numbers[:ncell]) & (cell_numbers[:ncell] < result.size)
+    result[cell_numbers[:ncell][valid]] = data[:ncell][valid]
+    return result
+
+
+def _adaptive_parent_owner_partition(coarse_mesh, parent_cell_numbers):
+    comm = coarse_mesh.comm
+    coarse_cell_numbers = _adaptive_netgen_cell_data(
+        coarse_mesh,
+        np.arange(adaptive_netgen_cell_count(coarse_mesh), dtype=IntType),
+    )
+    owned_cell_numbers = coarse_cell_numbers[:coarse_mesh.cell_set.size]
+    owned = np.empty((owned_cell_numbers.size, 2), dtype=IntType)
+    owned[:, 0] = owned_cell_numbers
+    owned[:, 1] = comm.rank
+    gathered = comm.gather(owned, root=0)
+
+    if comm.rank != 0:
+        return None, None
+
+    parent_owner = {}
+    for cells in gathered:
+        for cell_number, rank in cells:
+            parent_owner[int(cell_number)] = int(rank)
+
+    buckets = [[] for _ in range(comm.size)]
+    for fine_cell, parent_number in enumerate(parent_cell_numbers):
+        try:
+            owner = parent_owner[int(parent_number)]
+        except KeyError as exc:
+            raise RuntimeError(
+                "Cannot determine owner for adaptive parent cell "
+                f"{int(parent_number)}"
+            ) from exc
+        buckets[owner].append(fine_cell)
+
+    sizes = [len(bucket) for bucket in buckets]
+    points = np.asarray(
+        [point for bucket in buckets for point in bucket], dtype=IntType
+    )
+    return sizes, points
+
+
+def _set_adaptive_refined_mesh_metadata(mesh, netgen_mesh, netgen_flags,
+                                        distribution_parameters,
+                                        parent_cell_numbers):
+    mesh.netgen_mesh = netgen_mesh
+    mesh.netgen_flags = netgen_flags
+    mesh._distribution_parameters = dict(distribution_parameters)
+    mesh._adaptive_parent_cell_numbers = parent_cell_numbers
+    mesh._adaptive_parent_owned = True
+    mesh._adaptive_netgen_num_cells = len(adaptive_netgen_cells(mesh))
+
+
+def make_adaptive_refined_mesh(coarse_mesh, netgen_mesh, netgen_flags,
+                               parent_cell_numbers, redistribute=True):
+    from firedrake.mesh import Mesh
+
+    distribution_parameters = dict(coarse_mesh._distribution_parameters)
+    if coarse_mesh.comm.size == 1:
+        refined_mesh = Mesh(
+            netgen_mesh,
+            reorder=coarse_mesh._did_reordering,
+            distribution_parameters=distribution_parameters,
+            comm=coarse_mesh.comm,
+            netgen_flags=netgen_flags,
+            tolerance=coarse_mesh.tolerance,
+        )
+        _set_adaptive_refined_mesh_metadata(
+            refined_mesh, netgen_mesh, netgen_flags,
+            distribution_parameters, parent_cell_numbers,
+        )
+        return refined_mesh
+
+    parent_parameters = dict(distribution_parameters)
+    parent_parameters["partition"] = _adaptive_parent_owner_partition(
+        coarse_mesh, parent_cell_numbers
+    )
+
+    transfer_parameters = dict(parent_parameters)
+    transfer_parameters["overlap_type"] = (
+        DistributedMeshOverlapType.NONE, 0
+    )
+    transfer_mesh = Mesh(
+        netgen_mesh,
+        reorder=coarse_mesh._did_reordering,
+        distribution_parameters=transfer_parameters,
+        comm=coarse_mesh.comm,
+        netgen_flags=netgen_flags,
+        tolerance=coarse_mesh.tolerance,
+    )
+    _set_adaptive_refined_mesh_metadata(
+        transfer_mesh, netgen_mesh, netgen_flags,
+        distribution_parameters, parent_cell_numbers,
+    )
+
+    num_cells = transfer_mesh.cell_set.size
+    min_cells = transfer_mesh.comm.allreduce(num_cells, op=MPI.MIN)
+    max_cells = transfer_mesh.comm.allreduce(num_cells, op=MPI.MAX)
+    needs_redist = max_cells > 1.15 * min_cells
+
+    if not (redistribute and needs_redist):
+        overlap_type, overlap = distribution_parameters.get(
+            "overlap_type", (DistributedMeshOverlapType.FACET, 1)
+        )
+        if overlap_type == DistributedMeshOverlapType.NONE and overlap == 0:
+            return transfer_mesh
+        refined_mesh = Mesh(
+            netgen_mesh,
+            reorder=coarse_mesh._did_reordering,
+            distribution_parameters=parent_parameters,
+            comm=coarse_mesh.comm,
+            netgen_flags=netgen_flags,
+            tolerance=coarse_mesh.tolerance,
+        )
+        _set_adaptive_refined_mesh_metadata(
+            refined_mesh, netgen_mesh, netgen_flags,
+            distribution_parameters, parent_cell_numbers,
+        )
+        return refined_mesh
+
+    from firedrake.mg.utils import RedistMesh, redistribute_dm
+
+    redist_dm = transfer_mesh.topology_dm.clone()
+    redist_parameters = dict(distribution_parameters)
+    redist_parameters["partition"] = True
+    point_sf_orig, point_sf = redistribute_dm(redist_dm, redist_parameters)
+
+    redist_mesh_parameters = dict(redist_parameters)
+    redist_mesh_parameters["partition"] = False
+    redist_mesh_parameters["overlap_type"] = (
+        DistributedMeshOverlapType.NONE, 0
+    )
+    refined_mesh = Mesh(
+        redist_dm,
+        dim=coarse_mesh.geometric_dimension,
+        reorder=coarse_mesh._did_reordering,
+        distribution_parameters=redist_mesh_parameters,
+        comm=coarse_mesh.comm,
+        tolerance=coarse_mesh.tolerance,
+    )
+    _set_adaptive_refined_mesh_metadata(
+        refined_mesh, netgen_mesh, netgen_flags,
+        distribution_parameters, parent_cell_numbers,
+    )
+    if transfer_mesh.sfBC_orig is not None and point_sf_orig is not None:
+        refined_mesh.sfBC_orig = transfer_mesh.sfBC_orig.compose(point_sf_orig)
+        refined_mesh.sfBC = (refined_mesh.sfBC_orig if point_sf is point_sf_orig
+                             else transfer_mesh.sfBC_orig.compose(point_sf))
+    refined_mesh.redist = RedistMesh(transfer_mesh, refined_mesh, point_sf)
+    return refined_mesh
+
+
+def refine_marked_elements(mesh, mark, netgen_flags=None, redistribute=True):
+    if netgen_flags is None:
+        netgen_flags = mesh.netgen_flags
+    tdim = mesh.topological_dimension
+    if tdim not in {2, 3}:
+        raise NotImplementedError("No implementation for dimension other than 2 and 3.")
+    with mark.dat.vec as mvec:
+        if mesh.sfBC_orig is None:
+            cstart, cend = mesh.topology_dm.getHeightStratum(0)
+            cellNum = list(map(mesh._cell_numbering.getOffset, range(cstart, cend)))
+            mark_np = mvec.getArray()[cellNum]
+        else:
+            sfBCInv = mesh.sfBC_orig.createInverse()
+            _, mvec0 = mesh.topology_dm.distributeField(sfBCInv,
+                                                        mesh._cell_numbering,
+                                                        mvec)
+            mark_np = mvec0.getArray()
+    cells = (mesh.netgen_mesh.Elements3D() if tdim == 3
+             else mesh.netgen_mesh.Elements2D())
+    mark_np = mark_np[:len(cells)]
+    max_refs = 0 if mark_np.size == 0 else int(mark_np.max())
+    netgen_mesh = mesh.netgen_mesh.Copy()
+    parent_cell_numbers = np.arange(mark_np.size, dtype=IntType)
+    refine_faces = netgen_flags.get("refine_faces", False)
+    for r in range(max_refs):
+        cells = netgen_mesh.Elements3D() if tdim == 3 else netgen_mesh.Elements2D()
+        cells.NumPy()["refine"] = (mark_np[:len(cells)] > 0)
+        if tdim == 3:
+            faces = netgen_mesh.Elements2D()
+            faces.NumPy()["refine"] = refine_faces
+        netgen_mesh.Refine(adaptive=True)
+        mark_np -= 1
+        parents = netgen_mesh.parentelements if tdim == 3 else netgen_mesh.parentsurfaceelements
+        parents = parents.NumPy()["i"]
+        num_fine_cells = parents.shape[0]
+        num_coarse_cells = mark_np.size
+        indices = np.arange(num_fine_cells, dtype=IntType)
+        while (indices >= num_coarse_cells).any():
+            fine_cells = (indices >= num_coarse_cells)
+            indices[fine_cells] = parents[indices[fine_cells]]
+        parent_cell_numbers = parent_cell_numbers[indices]
+        if r < max_refs - 1:
+            mark_np = mark_np[indices]
+
+    return make_adaptive_refined_mesh(mesh, netgen_mesh, netgen_flags,
+                                      parent_cell_numbers, redistribute)
+
+
+def adaptive_cell_maps(coarse_mesh, fine_mesh, parent_cell_numbers):
+    if not (hasattr(coarse_mesh, "netgen_mesh") and hasattr(fine_mesh, "netgen_mesh")):
+        return None, None
+
+    coarse_cell_numbers = _adaptive_netgen_cell_data(
+        coarse_mesh, np.arange(adaptive_netgen_cell_count(coarse_mesh), dtype=IntType)
+    )
+    fine_parent_numbers = _adaptive_netgen_cell_data(
+        fine_mesh, np.asarray(parent_cell_numbers, dtype=IntType)
+    )
+
+    coarse_owned = coarse_mesh.cell_set.size
+    fine_owned = fine_mesh.cell_set.size
+
+    coarse_owned_lookup = {
+        int(cell_number): cell
+        for cell, cell_number in enumerate(coarse_cell_numbers[:coarse_owned])
+    }
+
+    fine_to_coarse = np.full((fine_owned, 1), -1, dtype=IntType)
+    for fine_cell, parent_number in enumerate(fine_parent_numbers[:fine_owned]):
+        fine_to_coarse[fine_cell, 0] = coarse_owned_lookup.get(int(parent_number), -1)
+
+    children = [[] for _ in range(coarse_owned)]
+    for fine_cell, parent_number in enumerate(fine_parent_numbers[:fine_owned]):
+        coarse_cell = coarse_owned_lookup.get(int(parent_number))
+        if coarse_cell is not None:
+            children[coarse_cell].append(fine_cell)
+
+    max_children = max((len(local_children) for local_children in children), default=0)
+    coarse_to_fine = np.full((coarse_owned, max(1, max_children)), -1, dtype=IntType)
+    for coarse_cell, local_children in enumerate(children):
+        coarse_to_fine[coarse_cell, :len(local_children)] = local_children
+
+    return coarse_to_fine, fine_to_coarse
+
+
+def adaptive_parent_cell_numbers(coarse_mesh, fine_mesh):
+    if not (hasattr(coarse_mesh, "netgen_mesh") and hasattr(fine_mesh, "netgen_mesh")):
+        return None
+
+    tdim = fine_mesh.topological_dimension
+    try:
+        parents = (fine_mesh.netgen_mesh.parentelements if tdim == 3
+                   else fine_mesh.netgen_mesh.parentsurfaceelements)
+    except AttributeError:
+        return None
+
+    parents = np.asarray(parents.NumPy()["i"], dtype=IntType)
+    nfine = len(adaptive_netgen_cells(fine_mesh))
+    ncoarse = adaptive_netgen_cell_count(coarse_mesh)
+    if parents.shape[0] < nfine or ncoarse == 0:
+        return None
+
+    indices = np.arange(nfine, dtype=IntType)
+    while True:
+        refined = indices >= ncoarse
+        if not refined.any():
+            break
+        parent_indices = parents[indices[refined]]
+        if (parent_indices < 0).any():
+            return None
+        indices[refined] = parent_indices
+    return indices
 
 
 @PETSc.Log.EventDecorator()
